Adjust order of evaluation of function arguments to match the spec
Adjust order of evaluation of function arguments to match the spec

The spec is very consistent about the order in which function arguments
are evaluated in cases where the function itself is not found, or is not
a function object. This PR changes that ordering to match the standard.

To implement this, the "lastStoredScriptable" in the Context is replaced
with a temporary object used to return state whenever a name or property
is looked up.

We also simplified the bytecode generation code for calling functions,
eliminating a lot of special cases that don't seem to help with
performance any more.

(cherry picked from commit 0f617f9af4c5428c33cc818e5221bea9316d1548)