Replace internal use of old "FunctionAndThis" methods
Replace usages of all the old "getXFunctionAndThis" functions with the
newer functions that return a LookupResult. This keeps us consistent
with the rest of the runtime. All of the old functions are deprecated and
could be removed if we can be sure that they're no longer used outside
this project.

(cherry picked from commit 5767c76d292b4e284fb1881ee3cb9d34907c4d05)